+++ py-polars/polars/interchange/__init__.py
@@ -1,0 +1,6 @@
+"""
+Module containing the implementation of the Python dataframe interchange protocol.
+
+Details on the protocol:
+https://data-apis.org/dataframe-protocol/latest/index.html
+"""
